==> jupyter/sales-russia/13-Sales_Russia-Analyse_Test_Train.ipynb <==
# importar pacotes necessários
import numpy as np
import pandas as pd

# definir parâmetros extras
pd.set_option('precision', 4)
pd.set_option('display.max_columns', 100)
pd.set_option('display.max_rows', 100)

input_prefix = 'data/'
#input_prefix = 'https://github.com/hjort/ai-labs/raw/master/jupyter/sales-russia/data/'

bzfile = '.bz2' # ''

df_test = pd.read_csv(
    input_prefix + 'test.csv' + bzfile, #nrows=10000,
    #dtype={'shop_id': 'category', 'item_id': 'category'},
    dtype={'shop_id': np.int8, 'item_id': np.int16},
    index_col='ID')
print('shape:', df_test.shape)
df_test.head()

df_test.info()

df_test.describe()

print('Total de linhas.....: %d' % df_test.shape[0])

shops_count = df_test['shop_id'].value_counts().count()
items_count = df_test['item_id'].value_counts().count()

print('Total de lojas......: %d' % shops_count)
print('Total de produtos...: %d' % items_count)
print('Total de combinações: %d (%d x %d)' % (shops_count * items_count, shops_count, items_count))

#dateparse = lambda x: pd.datetime.strptime(x, '%d.%m.%Y')
df_train = pd.read_csv(\
    input_prefix + 'sales_train.csv' + bzfile,
    #nrows=300, # somente primeiras linhas!
    #parse_dates=['date'], date_parser=dateparse,
    dtype={'date_block_num': np.int8, 'shop_id': np.int8, 'item_id': np.int16, \
           'item_price': np.float32, 'item_cnt_day': np.int32},
    )
print('shape:', df_train.shape)
df_train.head()

df_train2 = df_train.query('item_cnt_day > 0 and item_price > 0').drop(columns=['date'])
df_train2.head()

df_train3 = df_train2.groupby(['shop_id', 'item_id'], as_index=True).agg({
    'date_block_num': ['count', 'min', 'max'],
    'item_cnt_day': ['sum', 'min', 'max', 'mean', 'std', 'median'],
    'item_price': ['min', 'max', 'mean', 'std', 'median']
})

df_train3.head(20)

df_train3.info()

df_train3.describe()

# https://zaxrosenberg.com/pandas-multiindex-tutorial/

def flatten_cols(df: pd.DataFrame, delim: str = ""):
    """Flatten multiple column levels of the DataFrame into a one column level.

    Args:
        delim: the delimiter between the column values.

    Returns:
        A copy of the dataframe with the new column names.

    """
    new_cols = [delim.join((col_lev for col_lev in tup if col_lev))
                for tup in df.columns.values]
    ndf = df.copy()
    ndf.columns = new_cols

    return ndf

df_train4 = flatten_cols(df_train3, ":")
df_train4.head()

df_train4[['date_block_num:count', 'item_cnt_day:sum']].head()
