==> jupyter/sales-russia/49-prediction-variations.ipynb <==
import pandas as pd
import numpy as np

data = pd.read_csv('prediction-tbats.csv', index_col=['ID'])
data.head()

data.info()

data.describe()

data2 = data[['item_cnt_month']] / 2
data2.head()

data2.to_csv('prediction-tbats-div2.csv')

data3 = np.clip(data[['item_cnt_month']], 0, 10)
data3.head()

data3.describe()

data3.to_csv('prediction-tbats-clip10.csv')

data4 = np.ceil(data[['item_cnt_month']] / 2)
data4.head()

data4.to_csv('prediction-tbats-div2-ceil.csv')

data5 = np.floor(data[['item_cnt_month']] / 2)
data5.head()

data5.to_csv('prediction-tbats-div2-floor.csv')

data6 = np.ceil(data[['item_cnt_month']] / 3)
data6.head()

data6.to_csv('prediction-tbats-div3-ceil.csv')

data7 = np.sqrt(data[['item_cnt_month']])
data7.head()

data7.to_csv('prediction-tbats-sqrt.csv')

data8 = np.ceil(data[['item_cnt_month']] / 4)
data8.head()

data8.to_csv('prediction-tbats-div4-ceil.csv')

data9 = np.ceil(data[['item_cnt_month']] / 5)
data9.head()

data9.to_csv('prediction-tbats-div5-ceil.csv')

Your submission scored 1.21707, which is an improvement of your previous score of 1.21754. 

data10 = np.ceil(data[['item_cnt_month']] / 6)
data10.head()

data10.to_csv('prediction-tbats-div6-ceil.csv')

!gzip -9 prediction-tbats-*.csv; mv prediction-tbats-*.csv.gz submissions/

data11 = np.floor(data[['item_cnt_month']] / 6)
data11.head()

data11.to_csv('prediction-tbats-div6-floor.csv')

data12 = np.floor(data[['item_cnt_month']] / 5)
data12.head()

data12.to_csv('prediction-tbats-div5-floor.csv')

data13 = np.floor(data[['item_cnt_month']] / 4)
data13.head()

data13.to_csv('prediction-tbats-div4-floor.csv')

data14 = np.floor(data[['item_cnt_month']] / 3)
data14.head()

data14.to_csv('prediction-tbats-div3-floor.csv')

data15 = data[['item_cnt_month']] / 4
data15.head()

data15.to_csv('prediction-tbats-div4.csv')
